micro_service_with_gRPC/service_fastapi_gateway/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
numpy==2.0.2
pydantic
uvicorn
fastapi